검색 기능 추가중

+++ insider/src/main/java/com/kh/insider/repo/MemberRepo.java
@@ -15,6 +15,6 @@ public interface MemberRepo {
 	// 로그인 시각 갱신
 	public void updateLoginTime(long memberNo);
 	//단일 번호조회
-	public MemberDto findByNo(int memberNo);
+	public MemberDto findByNo(long memberNo);
 	public void changePassword(MemberDto memberDto);
 }
